Further refactor to fix behaviors of add and remove listener methods
We were using pass-by-value method receivers to methods responsible for
mutating the observer collection, so changed that to fix that. Also the
"Add" method was not checking for uniqueness of the item in a list, fixed
that too.

Error is no longer stored on the communication, rather returned from
methods and checked appropriately for responding thus.

Signed-off-by: Imran M Yousuf <imyousuf@newscred.com>