Make handlers receive specific events using the new Event and Packet API
Instead of handlers having to cast to specific events we can rather provide
the specific events in their specific handler and provide the client impls
to decide how to use them.

Also added an in memory packet registry to avoid having to handle the same
packet multiple times. Memory use would be:
session_count X packet_count X (9 bytes + golang internal use)
In theory it would not be more than a few megabytes in worst case in a
LAN scenario.

Signed-off-by: Imran M Yousuf <imyousuf@newscred.com>