Add internal listeners to add intelligence to communication impl
Basically when we receive a REGISTER message it is natural for the comm
layer itself to respond to the peer introducing itself. At the same time
we noticed that there is a need for listening to ping event to keep the
session-packet registry up to day. Hence we introduced an internal listener
that will hook to the communication API and process messages just like the
app would.

Signed-off-by: Imran M Yousuf <imyousuf@newscred.com>